Added GroupManager and SwitchServer functionality.
In its current form, this is definitely a CA-only patch.

GroupManager: Allows group-based network management policies to be enforced.
Specifically, a maximum number of distinct destinations per group can be
specified.  This functionality was specified by CA D2 as a requirement of
their initial IRIS rollout.

SwitchServer:  A module which allows for the control of a bank of Axis video
decoders.  Subject to GroupManager policies.